refactor: extract image loader and boot manager into dedicated modules
- Extract device path parsing and SFS handle matching from
  boot_services.rs into efi/image_loader.rs (~400 lines)
- Move boot manager dispatch logic from lib.rs into boot_manager.rs
  (~740 lines): run_boot_manager, try_boot_file_on_*, boot_selected_entry,
  boot_uefi_entry, boot_linux_entry, boot_payload_entry
- Keep store_device_globally() and with_disk() in lib.rs as they are
  used by both boot_manager.rs and menu.rs
- boot_services.rs: 2562 -> 2146 lines
- lib.rs: 1286 -> 567 lines
